some updates

+++ ElectricEngine.cpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-ElectricEngine::ElectricEngine(int power, int batteryCapacity)
+ElectricEngine::ElectricEngine(uint16_t power, uint16_t batteryCapacity)
     : Engine(power), batteryCapacity_(batteryCapacity) {
     std::cout << __FUNCTION__ << std::endl;
 }
